services/ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.34.74
kafka-python==2.0.2
pydantic==2.9.2
requests==2.32.3
structlog==24.1.0
jsonschema==4.23.0
python-dateutil==2.9.0.post0
pdfminer.six==20231228
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
prometheus-client==0.20.0
